Feat: authenticate by 42 email
- move directory mail to ft-auth
- authenticate with redis
