+++ src/ft-auth/ft-auth.controller.ts
@@ -1,20 +1,18 @@
-import { MailService } from './../mail/mail.service';
-import { Body, Controller, Post } from '@nestjs/common';
+import { Body, Controller, Post, Get, Query } from '@nestjs/common';
 import { FtAuthService } from './ft-auth.service';
-import { CreateFtAuthDto } from './dto/create-ft-auth.dto';
-import { UpdateAuthenticateDto } from './dto/update-authenticate.dto';
-import { Public } from '@root/auth/auth.decorator';
+import { GetUser } from '@root/auth/auth.decorator';
 
 @Controller('ft-auth')
 export class FtAuthController {
-  constructor(
-    private readonly ftAuthService: FtAuthService,
-    private readonly mailService: MailService,
-  ) {}
+  constructor(private readonly ftAuthService: FtAuthService) {}
 
-  @Public()
-  @Post()
-  sendMail(@Body('nickname') nickname: string) {
-    this.mailService.signin(nickname);
+  @Get()
+  sendMail(@GetUser('id') userId: number, @Query('nickname') nickname: string) {
+    this.ftAuthService.signin(nickname, userId);
+  }
+
+  @Get()
+  async getAuthCode(@Query('code') code: string) {
+    await this.ftAuthService.getAuth(code);
   }
 }
