+++ src/ft-auth/ft-auth.module.ts
@@ -1,4 +1,5 @@
-import { MailService } from './../mail/mail.service';
+import { UserModule } from '@user/user.module';
+import { UserService } from '@user/user.service';
 import { Module } from '@nestjs/common';
 import { FtAuthService } from './ft-auth.service';
 import { FtAuthController } from './ft-auth.controller';
@@ -6,8 +7,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { FtAuth } from './entities/ft-auth.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([FtAuth])],
+  imports: [TypeOrmModule.forFeature([FtAuth]), UserModule],
   controllers: [FtAuthController],
-  providers: [FtAuthService, MailService],
+  providers: [FtAuthService],
 })
 export class FtAuthModule {}
