+++ src/ft-auth/templates/signin.ejs
@@ -11,7 +11,7 @@
     "
   >
     <a
-      href="https://42world.com?code=<%= locals.code %>"
+      href="<%= locals.endpoint %>?code=<%= locals.code %>"
       style="color: white; font-size: 15px"
       >인증 하기</a
     >
