--- src/mail/mail.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { MailService } from './mail.service';
-import { MailController } from './mail.controller';
-
-@Module({
-  controllers: [MailController],
-  providers: [MailService],
-  exports: [MailService],
-})
-export class MailModule {}
